requirements/base.txt
async-timeout==4.0.2
asgiref==3.5.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
Deprecated==1.2.13
distro==1.6.0
Django==4.0.2
django-cors-headers
django-environ==0.8.1
django-filter==21.1
django-redis
djangorestframework==3.13.1
djangorestframework-gis==0.18
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-extra-fields
drf-yasg
factory-boy==3.2.1
Faker==13.11.0
gunicorn==20.1.0
idna==3.3
jsonschema==3.2.0
mccabe==0.6.1
packaging==21.3
paramiko==2.9.2
Pillow==9.0.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
redis==4.2.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
urllib3==1.26.8
whitenoise==6.2.0
websocket-client==0.59.0
wrapt==1.14.1

django-debug-toolbar==3.2.4